+++ SMBLibrary/Server/SMB2/CreateHelper.cs
@@ -121,8 +121,8 @@ namespace SMBLibrary.Server.SMB2
             response.LastWriteTime = entry.LastWriteTime;
             response.ChangeTime = entry.LastWriteTime;
             response.LastAccessTime = entry.LastAccessTime;
-            response.AllocationSize = NTFileSystemHelper.GetAllocationSize(entry.Size);
-            response.EndofFile = entry.Size;
+            response.AllocationSize = (long)NTFileSystemHelper.GetAllocationSize(entry.Size);
+            response.EndofFile = (long)entry.Size;
             return response;
         }
     }
